+++ lib/widgets/widget_emty_template.dart
@@ -5,7 +5,7 @@ class EmptyTemplate extends StatelessWidget {
   Widget build(BuildContext context) {
     return Container(
       height: MediaQuery.of(context).size.height,
-      width: MediaQuery.of(context).size.width * 0.22,
+      width: MediaQuery.of(context).size.width * 0.75,
       child: Card(
         elevation: 3,
         color: Colors.grey[100],