+++ CHANGELOG.md
@@ -44,6 +44,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Add invitation email on inexistent user
 - Add fingerprint authentication
 - Add wishlists categories
+- Add wishlist progression
 
 ### Changed